cleaned up proxy provider
